Lab 2
![the sudoku board](documentation_images/sudoku_board.png)

_Figure 1 - The sudoku board_

Below the game board, there will be another table consisting of every digit, except zero (`1`-`9`), as well as an undo icon (`images/undo.png`).

![the numbers palette](documentation_images/palette.png)

_Figure 2 - The numbers palette_

(`sudoku.css`)._
Additional Styling
When the cursor hovers over a cell in the Sudoku board, or in the numbers palette, it should change its background colour to `#f5dd90`.

Created two additional CSS classes, which will be used in the next part:
- `user-input` - will set the background colour to `#f2f2f2`
- `error` - will set the background colour to `#f76c5e`